Latest version

+++ Quest/libs/cocos2d/ccTypes.h
@@ -283,5 +283,5 @@ typedef struct _ccBlendFunc
 
 //! delta time type
 //! if you want more resolution redefine it as a double
-typedef float ccTime;
+typedef double ccTime;
 //typedef double ccTime;
